third_party/WebKit/LayoutTests/platform/win/media/video-enter-fullscreen-without-user-gesture-expected.txt

CONSOLE WARNING: line 14: Failed to execute 'requestFullscreen' on 'Element': API can only be initiated by a user gesture.
This is a testharness.js-based test.
PASS Test webkitRequestFullscreen() without any user gesture. 
Harness: the test ran to completion.
